Move clearing of entry filter from submission to filter change
- Fix filter for Select widget
- Filter out obsolete widgets and selected keys inside each widget
  filter
- Change logic of apply/clear visibility on lead filter
- Rename AppliedFilters to SourceAppliedFilters
- Break AppliedFilters into smaller components
