+++ app/utils/common.ts
@@ -133,7 +133,8 @@ export const enumKeySelector = <T>(d: EnumEntity<T>) => (
 );
 
 export const enumLabelSelector = <T extends string | number>(d: EnumEntity<T>) => (
-    capitalize(d.description ?? `${d.name}`)
+    // FIXME: why are we doing this
+    capitalize(d.description ?? String(d.name))
 );
 
 interface Options {
