+++ app/views/Project/Tagging/Sources/SourcesAppliedFilters/styles.css
@@ -1,0 +1,5 @@
+.applied-filters {
+    display: flex;
+    flex-wrap: wrap;
+    gap: var(--dui-spacing-medium);
+}
