+++ app/views/Project/Tagging/Sources/utils.tsx
@@ -1,15 +1,31 @@
-import { encodeDate } from '@togglecorp/fujs';
+import { encodeDate, listToMap, isDefined } from '@togglecorp/fujs';
 import { removeNull } from '@togglecorp/toggle-form';
-import { LeadOrderingEnum, ProjectSavedLeadFilterQuery } from '#generated/types';
+import {
+    // LeadOrderingEnum,
+    ProjectSavedLeadFilterQuery,
+} from '#generated/types';
 import { convertDateToIsoDateTime } from '#utils/common';
+import {
+    FrameworkFilterType,
+} from '#types/newAnalyticalFramework';
 import { FormType } from './SourcesFilter/schema';
 
 type RawSourcesFilter = NonNullable<NonNullable<NonNullable<ProjectSavedLeadFilterQuery>['project']>['userSavedLeadFilter']>['filters'];
 type FaramValues = Omit<FormType, 'projectId'>;
 
-enum SortDirection {
+// FIXME: move this to common utils
+// FIXME: get this from deep-ui
+export enum SortDirection {
     'asc' = 'Ascending',
     'dsc' = 'Descending',
+}
+
+// FIXME: move this to common utils
+function getDateString(dateTimeString: string | null | undefined) {
+    if (dateTimeString && Date.parse(dateTimeString)) {
+        return encodeDate(new Date(dateTimeString));
+    }
+    return dateTimeString;
 }
 
 export function transformSourcesFilterToEntriesFilter(filters: FaramValues) {
@@ -54,6 +70,7 @@ export function transformSourcesFilterToEntriesFilter(filters: FaramValues) {
     };
 }
 
+/*
 export function getSortState(
     orderingFilter: LeadOrderingEnum[] | null | undefined,
 ): { name: string; direction: SortDirection } | undefined {
@@ -63,15 +80,12 @@ export function getSortState(
         direction: direction === 'ASC' ? SortDirection.asc : SortDirection.dsc,
     }) : undefined;
 }
+*/
 
-function getDateString(dateTimeString: string | null | undefined) {
-    if (dateTimeString && Date.parse(dateTimeString)) {
-        return encodeDate(new Date(dateTimeString));
-    }
-    return dateTimeString;
-}
-
-export function transformRawFiltersToFormValues(filters: RawSourcesFilter) {
+export function transformRawFiltersToFormValues(
+    filters: RawSourcesFilter,
+    frameworkFilters: FrameworkFilterType[] | null | undefined,
+) {
     if (!filters) {
         return {};
     }
@@ -82,6 +96,13 @@ export function transformRawFiltersToFormValues(filters: RawSourcesFilter) {
         entriesFilterData,
         ...others
     } = filters;
+
+    const frameworkFiltersMapping = listToMap(
+        frameworkFilters ?? [],
+        (item) => item.key,
+        (item) => item,
+    );
+
     const formValues = {
         ...others,
         createdAt: getDateString(createdAt),
@@ -91,6 +112,33 @@ export function transformRawFiltersToFormValues(filters: RawSourcesFilter) {
             ...entriesFilterData,
             createdAtGte: getDateString(entriesFilterData?.createdAtGte),
             createdAtLte: getDateString(entriesFilterData?.createdAtLte),
+            filterableData: entriesFilterData?.filterableData?.map((item) => {
+                const key = item.filterKey;
+                const filterItem = frameworkFiltersMapping[key];
+                if (!filterItem) {
+                    return undefined;
+                }
+                if (
+                    filterItem.widgetType === 'SELECT'
+                    || filterItem.widgetType === 'MULTISELECT'
+                    || filterItem.widgetType === 'SCALE'
+                    || filterItem.widgetType === 'ORGANIGRAM'
+                    || filterItem.widgetType === 'MATRIX1D'
+                    || filterItem.widgetType === 'MATRIX2D'
+                ) {
+                    const keyMapping = listToMap(
+                        filterItem.properties?.options ?? [],
+                        (option) => option.key,
+                        () => true,
+                    );
+                    const newItem = {
+                        ...item,
+                        valueList: item.valueList?.filter((value) => keyMapping[value]),
+                    };
+                    return newItem;
+                }
+                return item;
+            }).filter(isDefined),
         },
     };
     return removeNull(formValues);
